Move common linting to workspace root
