+++ desktop/packages/mullvad-vpn/package.json
@@ -32,12 +32,10 @@
     "nseventmonitor": "^1.0.5"
   },
   "devDependencies": {
-    "@eslint/js": "^9.10.0",
     "@playwright/test": "^1.41.1",
     "@types/chai": "^4.3.3",
     "@types/chai-as-promised": "^7.1.5",
     "@types/chai-spies": "^1.0.3",
-    "@types/eslint__js": "^8.42.3",
     "@types/gettext-parser": "^4.0.1",
     "@types/google-protobuf": "^3.15.6",
     "@types/history": "^4.7.11",
@@ -57,13 +55,9 @@
     "electron": "30.0.4",
     "electron-builder": "^24.13.3",
     "electron-devtools-installer": "^3.2.0",
-    "eslint": "^9.10.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.36.1",
     "eslint-plugin-react-compiler": "^0.0.0-experimental-42acc6a-20241001",
     "eslint-plugin-react-hooks": "^0.0.0-experimental-2d16326d-20240930",
-    "eslint-plugin-simple-import-sort": "^12.1.1",
     "gettext-extractor": "^3.5.4",
     "globals": "^15.9.0",
     "grpc_tools_node_protoc_ts": "^5.3.2",
@@ -74,10 +68,8 @@
     "mocha": "^10.2.0",
     "playwright": "^1.41.1",
     "postject": "^1.0.0-alpha.6",
-    "prettier": "^3.3.3",
     "sinon": "^14.0.1",
     "tsc-watch": "^5.0.3",
-    "typescript-eslint": "^8.6.0",
     "vinyl-buffer": "^1.0.1",
     "vinyl-source-stream": "^2.0.0",
     "xvfb-maybe": "^0.2.1"
@@ -90,8 +82,7 @@
     "pack-test-executable": "./scripts/build-test-executable.sh",
     "build-test-executable": "npm run build && npm run pack-test-executable",
     "lint": "eslint .",
-    "format": "prettier \"**/*.{js,css,ts,tsx}\" --write",
-    "tsc": "tsc -p . --noEmit",
+    "lint-fix": "eslint --fix .",
     "e2e": "npm run build && npm run e2e:no-build",
     "e2e:no-build": "xvfb-maybe -- playwright test mocked",
     "e2e:sequential": "npm run build && npm run e2e:sequential:no-build",
