+++ desktop/packages/mullvad-vpn/tsconfig.json
@@ -17,7 +17,8 @@
     ],
     "typeRoots": [
       "./types",
-      "node_modules/@types"
+      "node_modules/@types",
+      "../../node_modules/@types"
     ]
   },
   "exclude": [
